Add dynamic imports for collection and portfolio pages to reduce initial bundle size
