+++ src/app/portfolio/page.tsx
@@ -1,32 +1,13 @@
-"use client";
+import dynamic from "next/dynamic";
+import { PageLoading } from "../../components/shared";
 
-import { PageLayout } from "../../components/layout";
-import {
-  Hero,
-  WhatIDo,
-  Experience,
-  Education,
-  TechnicalSkills,
-  FeaturedProjects,
-  AreasOfFocus,
-  SideProjects,
-  ContactSection,
-} from "../../components/portfolio";
+const PortfolioPageContent = dynamic(
+  () => import("../../components/portfolio/PortfolioPageContent"),
+  {
+    loading: () => <PageLoading />,
+  }
+);
 
 export default function PortfolioPage() {
-  return (
-    <PageLayout containerClassName="max-w-4xl">
-      <div className="mt-16 sm:mt-8">
-        <Hero />
-        <WhatIDo />
-        <TechnicalSkills />
-        <Experience />
-        <Education />
-        <FeaturedProjects />
-        <AreasOfFocus />
-        <SideProjects />
-        <ContactSection />
-      </div>
-    </PageLayout>
-  );
+  return <PortfolioPageContent />;
 }
