+++ src/components/collection/index.ts
@@ -1,6 +1,7 @@
 export { CollectionProvider, useCollection } from "../CollectionContext";
 export { default as CollectionControls } from "../CollectionControls";
 export { default as ReleaseGrid } from "../ReleaseGrid";
+export { default as CollectionPageContent } from "./CollectionPageContent";
 
 export type {
   Release,
